moved TipDectorDialog to dialogs/
