+++ neurolabi/gui/dialogs/tipdetectordialog.cpp
