+++ neurolabi/gui/dialogs/tipdetectordialog.h
