+++ neurolabi/gui/dialogs/tipdetectordialog.ui
